Move sync to promised wrapper to php-wasm/util
